classes/class-06-guardrails-ethics/demos/ai-agent-no-guardrails/requirements.txt
fastapi>=0.110
uvicorn>=0.29
python-dotenv>=1.0
httpx>=0.27
langgraph>=0.0.48
langchain-openai>=0.1.10
langchain-core>=0.1.52
